+++ src/theme/components/global.ts
@@ -42,6 +42,7 @@ const checkboxRadioHelperStyle = {
   ...helperTextMargin,
   marginLeft: "30px", // calc(22px + var(--nypl-space-xs))
   _disabled: {
+    color: "ui.gray.dark",
     fontStyle: "italic",
   },
 };
@@ -68,6 +69,12 @@ const labelLegendText = {
     fontWeight: "helper.default",
   },
 };
+// Used for the disabled state of Select and TextInput
+const selectTextinputDisabledStyles = {
+  borderColor: "ui.gray.light-cool",
+  color: "ui.gray.dark",
+  opacity: "1",
+}
 // Used for p, ul, and ol
 const textMargin = {
   margin: "0",
@@ -82,5 +89,6 @@ export {
   checkboxRadioLabelStyles,
   helperTextMargin,
   labelLegendText,
+  selectTextinputDisabledStyles,
   textMargin,
 };